requirements.txt
# ============================================================
# Runtime dependencies — installed on Heroku
# ============================================================

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12
jinja2>=3.1.4
aiofiles>=24.1.0

# LangChain core
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0

# Database / vector store
langchain-postgres>=0.0.12
psycopg2-binary>=2.9.9
pgvector>=0.3.3

# BM25 keyword retrieval
rank-bm25>=0.2.2

# Configuration & utilities
pyyaml>=6.0.2
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.6.0
httpx>=0.27.2
loguru>=0.7.2
itsdangerous>=2.2.0
